Moved NodeData to Data
