+++ WIP.md
@@ -2,6 +2,9 @@
 
 ## Current high-level goal
 
+- BUG: why does the flbrowser not connect directly to a CLI node sometimes?
+  - possibility: perhaps it waits for the CLI to connect, and doesn't initiate
+    a connection by itself?
 - send regular `NodeStat`s to the signalling-server
 - Update docker
 - Create usable crates for a signalling server and webrtc clients
@@ -16,17 +19,16 @@
 ### Things to do
 
 - CLEANUP:
-  - move `NodeData` into `Node`
   - rename `DataStorage` and perhaps move it back to flarch?
-- BUG: why does the flbrowser not connect directly to a CLI node sometimes?
-  - possibility: perhaps it waits for the CLI to connect, and doesn't initiate
-    a connection by itself?
+  - npm audit in flbrowser
 
 # Dates
 
 2022-05-24:
 - DISPLAY: flbrowser displays last ping as descending, but it should increase
 - BUG: messages get deleted after restart
+- CLEANUP:
+  - move `NodeData` into `Node`
 
 2022-05-22:
 - Finalize usage of broker-modules in `Node` and `Network`:
