+++ shared/flnode/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod node;
 pub mod version;
-pub mod node_data;
+// pub mod node_data;
 pub mod modules;
