+++ shared/flnode/tests/random_connections.rs
@@ -1,6 +1,6 @@
-use flnode::node_data::Brokers;
 use flarch::start_logging;
 mod helpers;
+use flnode::node::Brokers;
 use helpers::*;
 
 #[tokio::test]
@@ -31,8 +31,8 @@ async fn connect_nodes_n(nbr_nodes: usize) -> Result<(), NetworkError> {
         }
         nodes_visited.push(id);
         let node = net.nodes.get_mut(&id).unwrap();
-        node.node_data.update();
-        let nd = node.node_data.random.storage.clone();
+        node.node.update();
+        let nd = node.node.random.storage.clone();
         for n in nd.connected.0 {
             if nodes_to_visit.contains(&n.id) || nodes_visited.contains(&n.id) {
                 continue;
